Coordination proposal implementations
This is the first step to integrate the existing proposal structures with
the new mechanism. Here we are making them compliant with the
`coordinationProposal` interface and moving them out of the `chain.go` file
to indicate that they are no longer pure chain components.
